refactor(events): addon events in separate file
